Refactored MinPriorityQueue and Graph for better ergonomics
MinPirorityQueue can now be passed a simple f64 as cost, and
Graph::cost() now returns f64

+++ src/data_structures/graph.rs
@@ -3,7 +3,7 @@ use std::ops::Index;
 pub trait Graph<I, N>: Index<I, Output = N> { 
     fn neighbors(&self, _: &I) -> Vec<I>;
 
-    fn cost(&self, _: &I, _: &I) -> usize;
+    fn cost(&self, _: &I, _: &I) -> f64;
 
     fn contains(&self, _: &I) -> bool;
 }
